+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # @interop/vh-resource-log Changelog
 
-## 0.4.1 - TBD
+## 0.4.1 - 2026-08-22
 
 ### Fixed